fixed init location
